+++ apps/cli/src/commands/task/index.ts
@@ -1,0 +1,648 @@
+import type { Command } from 'commander';
+import pc from 'picocolors';
+
+import { getTrpcClient } from '../../api/client';
+import {
+  confirm,
+  displayWidth,
+  outputJson,
+  printTable,
+  timeAgo,
+  truncate,
+} from '../../utils/format';
+import { log } from '../../utils/logger';
+import { registerCheckpointCommands } from './checkpoint';
+import { registerDepCommands } from './dep';
+import { registerDocCommands } from './doc';
+import { briefIcon, priorityLabel, statusBadge } from './helpers';
+import { registerLifecycleCommands } from './lifecycle';
+import { registerReviewCommands } from './review';
+import { registerTopicCommands } from './topic';
+
+export function registerTaskCommand(program: Command) {
+  const task = program.command('task').description('Manage agent tasks');
+
+  // ── list ──────────────────────────────────────────────
+
+  task
+    .command('list')
+    .description('List tasks')
+    .option(
+      '--status <status>',
+      'Filter by status (pending/running/paused/completed/failed/canceled)',
+    )
+    .option('--root', 'Only show root tasks (no parent)')
+    .option('--parent <id>', 'Filter by parent task ID')
+    .option('--agent <id>', 'Filter by assignee agent ID')
+    .option('-L, --limit <n>', 'Page size', '50')
+    .option('--offset <n>', 'Offset', '0')
+    .option('--tree', 'Display as tree structure')
+    .option('--json [fields]', 'Output JSON')
+    .action(
+      async (options: {
+        agent?: string;
+        json?: string | boolean;
+        limit?: string;
+        offset?: string;
+        parent?: string;
+        root?: boolean;
+        status?: string;
+        tree?: boolean;
+      }) => {
+        const client = await getTrpcClient();
+
+        const input: Record<string, any> = {};
+        if (options.status) input.status = options.status;
+        if (options.root) input.parentTaskId = null;
+        if (options.parent) input.parentTaskId = options.parent;
+        if (options.agent) input.assigneeAgentId = options.agent;
+        if (options.limit) input.limit = Number.parseInt(options.limit, 10);
+        if (options.offset) input.offset = Number.parseInt(options.offset, 10);
+
+        // For tree mode, fetch all tasks (no pagination limit)
+        if (options.tree) {
+          input.limit = 100;
+          delete input.offset;
+        }
+
+        const result = await client.task.list.query(input as any);
+
+        if (options.json !== undefined) {
+          outputJson(result.data, options.json);
+          return;
+        }
+
+        if (!result.data || result.data.length === 0) {
+          log.info('No tasks found.');
+          return;
+        }
+
+        if (options.tree) {
+          // Build tree display
+          const taskMap = new Map<string, any>();
+          for (const t of result.data) taskMap.set(t.id, t);
+
+          const roots = result.data.filter((t: any) => !t.parentTaskId);
+          const children = new Map<string, any[]>();
+          for (const t of result.data) {
+            if (t.parentTaskId) {
+              const list = children.get(t.parentTaskId) || [];
+              list.push(t);
+              children.set(t.parentTaskId, list);
+            }
+          }
+
+          // Sort children by sortOrder first, then seq
+          for (const [, list] of children) {
+            list.sort(
+              (a: any, b: any) =>
+                (a.sortOrder ?? 0) - (b.sortOrder ?? 0) || (a.seq ?? 0) - (b.seq ?? 0),
+            );
+          }
+
+          const printNode = (t: any, prefix: string, isLast: boolean, isRoot: boolean) => {
+            const connector = isRoot ? '' : isLast ? '└── ' : '├── ';
+            const name = truncate(t.name || t.instruction, 40);
+            console.log(
+              `${prefix}${connector}${pc.dim(t.identifier)} ${statusBadge(t.status)} ${name}`,
+            );
+            const childList = children.get(t.id) || [];
+            const newPrefix = isRoot ? '' : prefix + (isLast ? '    ' : '│   ');
+            childList.forEach((child: any, i: number) => {
+              printNode(child, newPrefix, i === childList.length - 1, false);
+            });
+          };
+
+          for (const root of roots) {
+            printNode(root, '', true, true);
+          }
+          log.info(`Total: ${result.total}`);
+          return;
+        }
+
+        const rows = result.data.map((t: any) => [
+          pc.dim(t.identifier),
+          truncate(t.name || t.instruction, 40),
+          statusBadge(t.status),
+          priorityLabel(t.priority),
+          t.assigneeAgentId ? pc.dim(t.assigneeAgentId) : '-',
+          t.parentTaskId ? pc.dim('↳ subtask') : '',
+          timeAgo(t.createdAt),
+        ]);
+
+        printTable(rows, ['ID', 'NAME', 'STATUS', 'PRI', 'AGENT', 'TYPE', 'CREATED']);
+        log.info(`Total: ${result.total}`);
+      },
+    );
+
+  // ── view ──────────────────────────────────────────────
+
+  task
+    .command('view <id>')
+    .description('View task details (by ID or identifier like TASK-1)')
+    .option('--json [fields]', 'Output JSON')
+    .action(async (id: string, options: { json?: string | boolean }) => {
+      const client = await getTrpcClient();
+
+      const result = await client.task.detail.query({ id });
+
+      if (options.json !== undefined) {
+        outputJson(result.data, options.json);
+        return;
+      }
+
+      const t = result.data;
+
+      // ── Header ──
+      console.log(`\n${pc.bold(t.identifier)} ${t.name || ''}`);
+      console.log(
+        `${pc.dim('Status:')} ${statusBadge(t.status)}  ${pc.dim('Priority:')} ${priorityLabel(t.priority)}`,
+      );
+      console.log(`${pc.dim('Instruction:')} ${t.instruction}`);
+      if (t.description) console.log(`${pc.dim('Description:')} ${t.description}`);
+      if (t.assigneeAgentId) console.log(`${pc.dim('Agent:')} ${t.assigneeAgentId}`);
+      if (t.assigneeUserId) console.log(`${pc.dim('User:')} ${t.assigneeUserId}`);
+      if (t.parent) {
+        const p = t.parent as any;
+        console.log(`${pc.dim('Parent:')} ${p.identifier} ${p.name || ''}`);
+      }
+      console.log(
+        `${pc.dim('Topics:')} ${t.totalTopics}  ${pc.dim('Created:')} ${timeAgo(t.createdAt)}`,
+      );
+      if (t.heartbeatTimeout && t.lastHeartbeatAt) {
+        const hb = timeAgo(t.lastHeartbeatAt);
+        const interval = t.heartbeatInterval ? `${t.heartbeatInterval}s` : '-';
+        const elapsed = (Date.now() - new Date(t.lastHeartbeatAt).getTime()) / 1000;
+        const isStuck = t.status === 'running' && elapsed > t.heartbeatTimeout;
+        console.log(
+          `${pc.dim('Heartbeat:')} ${isStuck ? pc.red(hb) : hb}  ${pc.dim('interval:')} ${interval}  ${pc.dim('timeout:')} ${t.heartbeatTimeout}s${isStuck ? pc.red('  ⚠ TIMEOUT') : ''}`,
+        );
+      }
+      if (t.error) console.log(`${pc.red('Error:')} ${t.error}`);
+
+      // ── Subtasks ──
+      if (t.subtasks && t.subtasks.length > 0) {
+        // Build dependency lookup: taskId → [dependsOnIdentifier, ...]
+        const subtaskDeps = (t.subtaskDeps || []) as any[];
+        const idToIdentifier = new Map<string, string>();
+        for (const s of t.subtasks) idToIdentifier.set(s.id, s.identifier);
+
+        // Build lookup: which subtasks are completed
+        const completedIds = new Set(
+          t.subtasks.filter((s: any) => s.status === 'completed').map((s: any) => s.id),
+        );
+
+        const blockedBy = new Map<string, string[]>();
+        const hasUnresolvedDep = new Set<string>();
+        for (const d of subtaskDeps) {
+          if (d.type === 'blocks') {
+            const list = blockedBy.get(d.taskId) || [];
+            const depIdentifier = idToIdentifier.get(d.dependsOnId) || d.dependsOnId;
+            list.push(depIdentifier);
+            blockedBy.set(d.taskId, list);
+            // Check if the dependency is NOT completed
+            if (!completedIds.has(d.dependsOnId)) {
+              hasUnresolvedDep.add(d.taskId);
+            }
+          }
+        }
+
+        console.log(`\n${pc.bold('Subtasks:')}`);
+        for (const s of t.subtasks) {
+          const deps = blockedBy.get(s.id);
+          const depInfo = deps ? pc.dim(` ← blocks: ${deps.join(', ')}`) : '';
+          // Show 'blocked' instead of 'backlog' if task has unresolved dependencies
+          const displayStatus =
+            s.status === 'backlog' && hasUnresolvedDep.has(s.id) ? 'blocked' : s.status;
+          console.log(
+            `  ${pc.dim(s.identifier)} ${statusBadge(displayStatus)} ${s.name || s.instruction}${depInfo}`,
+          );
+        }
+      }
+
+      // ── Dependencies ──
+      if (t.dependencies && t.dependencies.length > 0) {
+        console.log(`\n${pc.bold('Dependencies:')}`);
+        for (const d of t.dependencies as any[]) {
+          console.log(`  ${pc.dim(d.type)}: ${d.dependsOnId}`);
+        }
+      }
+
+      // ── Checkpoint ──
+      {
+        const cp = t.checkpoint as any;
+        console.log(`\n${pc.bold('Checkpoint:')}`);
+        const hasConfig =
+          cp.onAgentRequest !== undefined ||
+          cp.topic?.before ||
+          cp.topic?.after ||
+          cp.tasks?.beforeIds?.length > 0 ||
+          cp.tasks?.afterIds?.length > 0;
+
+        if (hasConfig) {
+          if (cp.onAgentRequest !== undefined)
+            console.log(`  onAgentRequest: ${cp.onAgentRequest}`);
+          if (cp.topic?.before) console.log(`  topic.before: ${cp.topic.before}`);
+          if (cp.topic?.after) console.log(`  topic.after: ${cp.topic.after}`);
+          if (cp.tasks?.beforeIds?.length > 0)
+            console.log(`  tasks.before: ${cp.tasks.beforeIds.join(', ')}`);
+          if (cp.tasks?.afterIds?.length > 0)
+            console.log(`  tasks.after: ${cp.tasks.afterIds.join(', ')}`);
+        } else {
+          console.log(`  ${pc.dim('(not configured, default: onAgentRequest=true)')}`);
+        }
+      }
+
+      // ── Review ──
+      {
+        const rv = t.review as any;
+        console.log(`\n${pc.bold('Review:')}`);
+        if (rv && rv.enabled) {
+          console.log(
+            `  judge: ${rv.judge?.model || 'default'}${rv.judge?.provider ? ` (${rv.judge.provider})` : ''}`,
+          );
+          console.log(`  maxIterations: ${rv.maxIterations}  autoRetry: ${rv.autoRetry}`);
+          if (rv.rubrics?.length > 0) {
+            for (let i = 0; i < rv.rubrics.length; i++) {
+              const rb = rv.rubrics[i];
+              const threshold = rb.threshold ? ` ≥ ${Math.round(rb.threshold * 100)}%` : '';
+              const typeTag = pc.dim(`[${rb.type}]`);
+              let configInfo = '';
+              if (rb.type === 'llm-rubric') configInfo = rb.config?.criteria || '';
+              else if (rb.type === 'contains' || rb.type === 'equals')
+                configInfo = `value="${rb.config?.value}"`;
+              else if (rb.type === 'regex') configInfo = `pattern="${rb.config?.pattern}"`;
+              console.log(`  ${i + 1}. ${rb.name} ${typeTag}${threshold} ${pc.dim(configInfo)}`);
+            }
+          }
+        } else {
+          console.log(`  ${pc.dim('(not configured)')}`);
+        }
+      }
+
+      // ── Activities (last section) ──
+      {
+        // ── Workspace ──
+        {
+          const workspace = t.documents as any;
+          const nodeMap = workspace?.nodeMap || {};
+          const tree = workspace?.tree || [];
+          const docCount = Object.keys(nodeMap).length;
+
+          if (docCount === 0) {
+            console.log(`\n${pc.bold('Workspace:')}`);
+            console.log(`  ${pc.dim('No documents yet.')}`);
+          } else {
+            console.log(`\n${pc.bold(`Workspace (${docCount}):`)}`);
+
+            const formatSize = (chars: number | null) => {
+              if (!chars) return '';
+              if (chars >= 10_000) return `${(chars / 1000).toFixed(1)}k`;
+              return `${chars}`;
+            };
+
+            const LEFT_COL = 56; // fixed display width from start to after title (including indent)
+            const FROM_WIDTH = 10;
+
+            const printNode = (node: any, prefix: string, isLast: boolean, isRoot: boolean) => {
+              const doc = nodeMap[node.id];
+              if (!doc) return;
+              const isFolder = doc.fileType === 'folder' || doc.fileType === 'custom/folder';
+              const icon = isFolder ? '📁' : '📄';
+              const connector = isRoot ? '  ' : isLast ? '  └── ' : '  ├── ';
+
+              // Prefix + connector + icon take variable width; pad title to fill LEFT_COL
+              const leftPrefix = `${prefix}${connector}${icon} `;
+              const prefixWidth = displayWidth(leftPrefix);
+              const maxTitle = Math.max(10, LEFT_COL - prefixWidth);
+              const titleStr = truncate(doc.title, maxTitle);
+              const titleWidth = displayWidth(titleStr);
+              const titlePad = ' '.repeat(Math.max(1, LEFT_COL - prefixWidth - titleWidth));
+
+              // Source column (fixed width)
+              const fromStr = doc.sourceTaskIdentifier ? `← ${doc.sourceTaskIdentifier}` : '';
+              const fromPad = ' '.repeat(Math.max(1, FROM_WIDTH - fromStr.length + 1));
+
+              // Right columns
+              const size =
+                !isFolder && doc.charCount
+                  ? formatSize(doc.charCount).padStart(6) + ' chars'
+                  : ''.padStart(12);
+              const updated = doc.updatedAt ? timeAgo(doc.updatedAt).padStart(7) : '';
+
+              console.log(
+                `${leftPrefix}${titleStr}${titlePad}${pc.dim(`(${node.id})`)}  ${fromStr}${fromPad}${pc.dim(`${size}  ${updated}`)}`,
+              );
+
+              const kids = node.children || [];
+              const childPrefix = isRoot ? '' : prefix + (isLast ? '      ' : '  │   ');
+              kids.forEach((child: any, i: number) => {
+                printNode(child, childPrefix, i === kids.length - 1, false);
+              });
+            };
+
+            tree.forEach((node: any, i: number) => {
+              printNode(node, '', i === tree.length - 1, true);
+            });
+          }
+        }
+
+        const activities: {
+          data: any;
+          time: number;
+          type: 'topic' | 'brief' | 'comment' | 'review';
+        }[] = [];
+
+        for (const tp of t.topics || []) {
+          activities.push({
+            data: tp,
+            time: new Date(tp.createdAt).getTime(),
+            type: 'topic',
+          });
+          // Add review as separate activity if topic has been reviewed
+          if (tp.reviewedAt) {
+            activities.push({
+              data: tp,
+              time: new Date(tp.reviewedAt).getTime(),
+              type: 'review',
+            });
+          }
+        }
+
+        for (const b of t.briefs || []) {
+          activities.push({
+            data: b,
+            time: new Date(b.createdAt).getTime(),
+            type: 'brief',
+          });
+        }
+
+        for (const c of t.comments || []) {
+          activities.push({
+            data: c,
+            time: new Date(c.createdAt).getTime(),
+            type: 'comment',
+          });
+        }
+
+        console.log(`\n${pc.bold('Activities:')}`);
+        if (activities.length === 0) {
+          console.log(`  ${pc.dim('No activities yet.')}`);
+        } else {
+          activities.sort((a, b) => a.time - b.time);
+
+          const pad = (s: string, w: number) => s.padStart(w);
+
+          for (const act of activities) {
+            const ago = pad(
+              timeAgo(act.type === 'review' ? act.data.reviewedAt : act.data.createdAt),
+              7,
+            );
+
+            if (act.type === 'topic') {
+              const tp = act.data;
+              const sBadge = statusBadge(tp.status || 'running');
+              console.log(
+                `  💬 ${pc.dim(ago)} Topic #${tp.seq} ${tp.title || 'Untitled'} ${sBadge}  ${pc.dim(tp.id)}`,
+              );
+            } else if (act.type === 'review') {
+              const tp = act.data;
+              const passed = tp.reviewPassed === 1;
+              const icon = passed ? pc.green('✓') : pc.red('✗');
+              const scoreText = ((tp.reviewScores as any[]) || [])
+                .map((s: any) => {
+                  const pct = Math.round(s.score * 100);
+                  const sIcon = s.passed ? pc.green('✓') : pc.red('✗');
+                  return `${s.rubricId} ${pct}%${sIcon}`;
+                })
+                .join(' | ');
+              const iter =
+                tp.reviewIteration > 1 ? pc.dim(` (iteration ${tp.reviewIteration})`) : '';
+              console.log(
+                `  🔍 ${pc.dim(ago)} Review ${icon} ${passed ? 'passed' : 'failed'} (${tp.reviewScore}%) ${pc.dim(scoreText)}${iter}`,
+              );
+            } else if (act.type === 'comment') {
+              const c = act.data;
+              const author = c.agentId ? `🤖 ${c.agentId}` : '👤 user';
+              console.log(`  💭 ${pc.dim(ago)} ${pc.cyan(author)} ${c.content}`);
+            } else {
+              const b = act.data;
+              const icon = briefIcon(b.type);
+              const pri =
+                b.priority === 'urgent'
+                  ? pc.red(' [urgent]')
+                  : b.priority === 'normal'
+                    ? pc.yellow(' [normal]')
+                    : '';
+              let statusTag: string;
+              if (b.resolvedAt) {
+                const actions = (b.actions as any[]) || [];
+                const matched = actions.find((a: any) => a.key === b.resolvedAction);
+                statusTag = pc.green(` ${matched?.label || '✓'}`);
+              } else if (b.readAt) {
+                statusTag = pc.dim(' (read)');
+              } else {
+                statusTag = pc.yellow(' ●');
+              }
+              const typeLabel = pc.dim(`[${b.type}]`);
+              console.log(
+                `  ${icon} ${pc.dim(ago)} Brief ${typeLabel} ${b.title}${pri}${statusTag}  ${pc.dim(b.id)}`,
+              );
+            }
+          }
+        }
+      }
+
+      console.log();
+    });
+
+  // ── create ──────────────────────────────────────────────
+
+  task
+    .command('create')
+    .description('Create a new task')
+    .requiredOption('-i, --instruction <text>', 'Task instruction')
+    .option('-n, --name <name>', 'Task name')
+    .option('--agent <id>', 'Assign to agent')
+    .option('--parent <id>', 'Parent task ID')
+    .option('--priority <n>', 'Priority (0=none, 1=urgent, 2=high, 3=normal, 4=low)', '0')
+    .option('--prefix <prefix>', 'Identifier prefix', 'TASK')
+    .option('--json [fields]', 'Output JSON')
+    .action(
+      async (options: {
+        agent?: string;
+        instruction: string;
+        json?: string | boolean;
+        name?: string;
+        parent?: string;
+        prefix?: string;
+        priority?: string;
+      }) => {
+        const client = await getTrpcClient();
+
+        const input: Record<string, any> = {
+          instruction: options.instruction,
+        };
+        if (options.name) input.name = options.name;
+        if (options.agent) input.assigneeAgentId = options.agent;
+        if (options.parent) input.parentTaskId = options.parent;
+        if (options.priority) input.priority = Number.parseInt(options.priority, 10);
+        if (options.prefix) input.identifierPrefix = options.prefix;
+
+        const result = await client.task.create.mutate(input as any);
+
+        if (options.json !== undefined) {
+          outputJson(result.data, options.json);
+          return;
+        }
+
+        log.info(`Task created: ${pc.bold(result.data.identifier)} ${result.data.name || ''}`);
+      },
+    );
+
+  // ── edit ──────────────────────────────────────────────
+
+  task
+    .command('edit <id>')
+    .description('Update a task')
+    .option('-n, --name <name>', 'Task name')
+    .option('-i, --instruction <text>', 'Task instruction')
+    .option('--agent <id>', 'Assign to agent')
+    .option('--priority <n>', 'Priority (0-4)')
+    .option('--heartbeat-interval <n>', 'Heartbeat interval in seconds')
+    .option('--heartbeat-timeout <n>', 'Heartbeat timeout in seconds (0 to disable)')
+    .option('--description <text>', 'Task description')
+    .option('--json [fields]', 'Output JSON')
+    .action(
+      async (
+        id: string,
+        options: {
+          agent?: string;
+          description?: string;
+          heartbeatInterval?: string;
+          heartbeatTimeout?: string;
+          instruction?: string;
+          json?: string | boolean;
+          name?: string;
+          priority?: string;
+        },
+      ) => {
+        const client = await getTrpcClient();
+
+        const input: Record<string, any> = { id };
+        if (options.name) input.name = options.name;
+        if (options.instruction) input.instruction = options.instruction;
+        if (options.description) input.description = options.description;
+        if (options.agent) input.assigneeAgentId = options.agent;
+        if (options.priority) input.priority = Number.parseInt(options.priority, 10);
+        if (options.heartbeatInterval)
+          input.heartbeatInterval = Number.parseInt(options.heartbeatInterval, 10);
+        if (options.heartbeatTimeout !== undefined) {
+          const val = Number.parseInt(options.heartbeatTimeout, 10);
+          input.heartbeatTimeout = val === 0 ? null : val;
+        }
+
+        const result = await client.task.update.mutate(input as any);
+
+        if (options.json !== undefined) {
+          outputJson(result.data, typeof options.json === 'string' ? options.json : undefined);
+          return;
+        }
+
+        log.info(`Task updated: ${pc.bold(result.data.identifier)}`);
+      },
+    );
+
+  // ── delete ──────────────────────────────────────────────
+
+  task
+    .command('delete <id>')
+    .description('Delete a task')
+    .option('-y, --yes', 'Skip confirmation')
+    .action(async (id: string, options: { yes?: boolean }) => {
+      if (!options.yes) {
+        const ok = await confirm(`Delete task ${pc.bold(id)}?`);
+        if (!ok) return;
+      }
+
+      const client = await getTrpcClient();
+      await client.task.delete.mutate({ id });
+      log.info(`Task ${pc.bold(id)} deleted.`);
+    });
+
+  // ── clear ──────────────────────────────────────────────
+
+  task
+    .command('clear')
+    .description('Delete all tasks')
+    .option('-y, --yes', 'Skip confirmation')
+    .action(async (options: { yes?: boolean }) => {
+      if (!options.yes) {
+        const ok = await confirm(`Delete ${pc.red('ALL')} tasks? This cannot be undone.`);
+        if (!ok) return;
+      }
+
+      const client = await getTrpcClient();
+      const result = (await client.task.clearAll.mutate()) as any;
+      log.info(`${result.count} task(s) deleted.`);
+    });
+
+  // ── tree ──────────────────────────────────────────────
+
+  task
+    .command('tree <id>')
+    .description('Show task tree (subtasks + dependencies)')
+    .option('--json [fields]', 'Output JSON')
+    .action(async (id: string, options: { json?: string | boolean }) => {
+      const client = await getTrpcClient();
+      const result = await client.task.getTaskTree.query({ id });
+
+      if (options.json !== undefined) {
+        outputJson(result.data, options.json);
+        return;
+      }
+
+      if (!result.data || result.data.length === 0) {
+        log.info('No tasks found.');
+        return;
+      }
+
+      // Build tree display (raw SQL returns snake_case)
+      const taskMap = new Map<string, any>();
+      for (const t of result.data) taskMap.set(t.id, t);
+
+      const printNode = (taskId: string, indent: number) => {
+        const t = taskMap.get(taskId);
+        if (!t) return;
+
+        const prefix = indent === 0 ? '' : '  '.repeat(indent) + '├── ';
+        const name = t.name || t.identifier || '';
+        const status = t.status || 'pending';
+        const identifier = t.identifier || t.id;
+        console.log(`${prefix}${pc.dim(identifier)} ${statusBadge(status)} ${name}`);
+
+        // Print children (handle both camelCase and snake_case)
+        for (const child of result.data) {
+          const childParent = child.parentTaskId || child.parent_task_id;
+          if (childParent === taskId) {
+            printNode(child.id, indent + 1);
+          }
+        }
+      };
+
+      // Find root - resolve identifier first
+      const resolved = await client.task.find.query({ id });
+      const rootId = resolved.data.id;
+      const root = result.data.find((t: any) => t.id === rootId);
+      if (root) printNode(root.id, 0);
+      else log.info('Root task not found in tree.');
+    });
+
+  // Register subcommand groups
+  registerLifecycleCommands(task);
+  registerCheckpointCommands(task);
+  registerReviewCommands(task);
+  registerDepCommands(task);
+  registerTopicCommands(task);
+  registerDocCommands(task);
+}
