+++ packages/types/src/task/index.ts
@@ -9,3 +9,24 @@ export interface CheckpointConfig {
     before?: boolean;
   };
 }
+
+export interface WorkspaceDocNode {
+  charCount: number | null;
+  createdAt: string;
+  fileType: string;
+  parentId: string | null;
+  pinnedBy: string;
+  sourceTaskIdentifier: string | null;
+  title: string;
+  updatedAt: string | null;
+}
+
+export interface WorkspaceTreeNode {
+  children: WorkspaceTreeNode[];
+  id: string;
+}
+
+export interface WorkspaceData {
+  nodeMap: Record<string, WorkspaceDocNode>;
+  tree: WorkspaceTreeNode[];
+}
