+++ src/server/routers/lambda/document.ts
@@ -27,7 +27,7 @@ export const documentRouter = router({
     .input(
       z.object({
         content: z.string().optional(),
-        editorData: z.string(),
+        editorData: z.string().optional(),
         fileType: z.string().optional(),
         knowledgeBaseId: z.string().optional(),
         metadata: z.record(z.any()).optional(),
@@ -47,7 +47,7 @@ export const documentRouter = router({
       }
 
       // Parse editorData from JSON string to object
-      const editorData = JSON.parse(input.editorData);
+      const editorData = input.editorData ? JSON.parse(input.editorData) : undefined;
       return ctx.documentService.createDocument({
         ...input,
         editorData,
